+++ reclazztest/src/com/onurkat/reclazztest/services/TestStatus.java
@@ -1,0 +1,9 @@
+/*
+ * Copyright 2026 Onur Kat
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package com.onurkat.reclazztest.services;
+
+public enum TestStatus {
+    OPEN, CLOSED, ARCHIVED
+}